[#14] layout component props로 전달
